pyproject.toml
[project]
name = "omni-agent"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.40.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

